Fix real dtype mismatch in the new backward kernels: cast bf16 loads before dotting with fp32 accumulators
Real bug from the first CUDA run of the 3 new backward kernels: all 5
correctness cases failed with "AssertionError: Both operands must be
same dtype. Got fp32 and bf16" during Triton compilation. Root cause:
the forward kernel's own working pattern explicitly casts a freshly-
loaded bf16 tensor to fp32 before dotting it against an fp32 accumulator
(v = tl.load(...).to(tl.float32) before acc += tl.dot(scores, v, ...)) --
I missed that same cast on the second tl.dot in each of the 3 new
kernels: q_b in _bdh_dq_query_role_kernel, q_a in
_bdh_dq_key_role_kernel, and dout_a in _bdh_dv_kernel, each multiplied
against an fp32 dscore/score accumulator without first casting the
freshly bf16-loaded operand. Added .to(tl.float32) at all 3 sites,
matching the forward kernel's already-proven pattern exactly. The first
tl.dot in each kernel (both operands freshly loaded bf16, e.g. dout/v or
q_b/q_a) needed no change -- that combination already matches the
forward kernel's own first dot and was never the problem.

Full local suite still green, 784 passed. Not yet re-verified on CUDA --
re-dispatching.

+++ reference/hz0h_bdh_triton_attention_torch.py
@@ -220,7 +220,7 @@ if _HAS_TRITON:
             q_b = tl.load(
                 q_ptr + b_rows[:, None] * q_stride_t + n_out_cols[None, :] * q_stride_n,
                 mask=b_mask[:, None] & n_out_mask[None, :], other=0.0,
-            )
+            ).to(tl.float32)
             acc += tl.dot(dscore, q_b, input_precision="ieee")
 
         tl.store(
@@ -278,7 +278,7 @@ if _HAS_TRITON:
             q_a = tl.load(
                 q_ptr + a_rows[:, None] * q_stride_t + n_out_cols[None, :] * q_stride_n,
                 mask=a_mask[:, None] & n_out_mask[None, :], other=0.0,
-            )
+            ).to(tl.float32)
             acc += tl.dot(dscore, q_a, input_precision="ieee")
 
         tl.store(
@@ -333,7 +333,7 @@ if _HAS_TRITON:
             dout_a = tl.load(
                 dout_ptr + a_rows[:, None] * dout_stride_t + d_cols[None, :] * dout_stride_d,
                 mask=a_mask[:, None] & d_mask[None, :], other=0.0,
-            )
+            ).to(tl.float32)
             acc += tl.dot(score, dout_a, input_precision="ieee")
 
         tl.store(
